Dungeon terrain despawn

+++ src/Application/include/level/LevelTilemapBuilder.hpp
@@ -17,7 +17,8 @@ class TilemapBuilder {
 public:
 
     TilemapBuilder(engine::Shader *shader, glm::ivec2 &&size = { 100, 100 }) :
-        m_shader(shader), m_size(size), m_tiles(static_cast<std::size_t>(m_size.x * m_size.y), TileEnum::NONE)
+        m_shader(shader), m_size(size),
+        m_tiles(static_cast<std::size_t>(m_size.x) * static_cast<std::size_t>(m_size.y), TileEnum::NONE)
     {
     }
